% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_alphamis_txf}
\alias{get_alphamis_txf}
\title{obtain AlphaMissense TabixFile from cache, after
retrieving the relevant resource if not present there}
\usage{
get_alphamis_txf(cache = BiocFileCache::BiocFileCache(), build = "hg38")
}
\arguments{
\item{cache}{An instance of BiocFileCache}

\item{build}{character(1) defaults to "hg38"}
}
\description{
obtain AlphaMissense TabixFile from cache, after
retrieving the relevant resource if not present there
}
\examples{
ca = BiocFileCache::BiocFileCache()
txf = get_alphamis_txf(ca)
Rsamtools::yieldSize(txf) = 10L
read.delim(text=Rsamtools::scanTabix(txf)[[1]], h=FALSE) 
}
